+++ app/server-binance.js
@@ -1,8 +1,7 @@
 const _ = require('lodash');
 // const moment = require('moment-timezone');
 const config = require('config');
-const moment = require('moment');
-const { PubSub, cache, mongo, slack } = require('./helpers');
+const { PubSub, cache, mongo } = require('./helpers');
 
 const { maskConfig } = require('./cronjob/trailingTradeHelper/util');
 const {
@@ -12,8 +11,7 @@ const {
   lockSymbol,
   unlockSymbol,
   getAccountInfoFromAPI,
-  cacheExchangeSymbols,
-  getAPILimit
+  cacheExchangeSymbols
 } = require('./cronjob/trailingTradeHelper/common');
 const {
   getWebsocketATHCandlesClean,
@@ -183,41 +181,7 @@ const runBinance = async serverLogger => {
     `Binance ${config.get('mode')} started on`
   );
 
-  try {
-    await setupBinance(logger);
-  } catch (err) {
-    // For the redlock fail
-    if (err.message.includes('redlock')) {
-      // Simply ignore
-      return;
-    }
-
-    logger.error(
-      { err, errorCode: err.code, debug: true },
-      `⚠ Setup binance failed.`
-    );
-
-    if (
-      err.code === -1001 ||
-      err.code === -1021 || // Timestamp for this request is outside the recvWindow
-      err.code === 'ECONNRESET' ||
-      err.code === 'ECONNREFUSED'
-    ) {
-      // Let's silent for internal server error or assumed temporary errors
-    } else {
-      slack.sendMessage(
-        `Setup binance failed (${moment().format('HH:mm:ss.SSS')})\n` +
-          `Code: ${err.code}\n` +
-          `Message:\`\`\`${err.message}\`\`\`\n` +
-          `${
-            config.get('featureToggle.notifyDebug')
-              ? `Stack:\`\`\`${err.stack}\`\`\`\n`
-              : ''
-          }` +
-          `- Current API Usage: ${getAPILimit(logger)}`
-      );
-    }
-  }
+  await setupBinance(logger);
 };
 
 module.exports = { runBinance };