+++ app/server.js
@@ -2,11 +2,15 @@ const { logger: rootLogger, mongo } = require('./helpers');
 const { runBinance } = require('./server-binance');
 const { runCronjob } = require('./server-cronjob');
 const { runFrontend } = require('./server-frontend');
+const { runErrorHandler } = require('./error-handler');
 
 (async () => {
   const logger = rootLogger.child({
     gitHash: process.env.GIT_HASH || 'unspecified'
   });
+
+  runErrorHandler(logger);
+
   await mongo.connect(logger);
 
   await runBinance(logger);