+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+- Refactored error handling - [#434](https://github.com/chrisleekr/binance-trading-bot/pull/434)
+
 ## [0.0.86] - 2022-07-22
 
 - Upgraded TradingView TA to 3.2.10 - [#426](https://github.com/chrisleekr/binance-trading-bot/pull/426)